fix the teacher role tasks for managing his own articles

+++ app/views/Dashboard/pages/courses.php
@@ -1,7 +1,7 @@
 <?php
 
 require_once __DIR__ . '/../../../../vendor/autoload.php';
-// require_once __DIR__ . '/../../../Config/error_config.php';
+require_once __DIR__ . '/../../../Config/error_config.php';
 
 use App\Classes\Session;
 use App\Classes\BaseModel;
@@ -11,16 +11,33 @@ use App\Classes\User;
 use App\Controllers\CourseController;
 
 Session::start();
+
+if(!Session::exists('user')) {
+    header("Location: ../../../public/index.php");
+    exit;
+} 
+
+$role = Session::get('user')[0]['role'];
+$userId = Session::get('user')[0]['id'];
 $baseModel = new BaseModel;
 
 $catgory = new Category;
 
 $courseContr = new CourseController;
-$courses = $courseContr->getAllCourses();
 
 echo "<pre>";
 // var_dump($_SESSION);
 echo "</pre>";
+
+if ($role === 'admin') {
+    $courses = $courseContr->getAllCourses();
+} else if ($role === 'teacher') {
+    $courses = $courseContr->getTeacherCourses($userId);
+}
+
+if (isset($_POST['delete-course']) && $_SERVER['REQUEST_METHOD'] === 'POST') {
+    $courseContr->deleteCourse($_POST['course-id']);
+}
 
 ?>
 
@@ -35,11 +52,37 @@ echo "</pre>";
     <meta
         name="description"
         content="Youdemy - Online Learning Platform" />
-    <?php include '../../partials/head.php' ?>
+    <?php include '../../components/head.php' ?>
     <link rel="stylesheet" href="../../../node_modules/apexcharts/dist/apexcharts.css" />
     <link rel="stylesheet" href="../../../public/assets/css/theme.css" />
     <link rel="stylesheet" href="../../../public/assets/css/styles.css" />
     <script src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js" defer></script>
+
+    <style>
+        /* Custom CSS for modal positioning */
+        .modal-overlay {
+            position: fixed;
+            top: 0;
+            left: 0;
+            right: 0;
+            bottom: 0;
+            background-color: rgba(0, 0, 0, 0.5);
+            display: none;
+            align-items: center;
+            justify-content: center;
+            z-index: 1000;
+        }
+
+        .modal-content {
+            position: relative;
+            background-color: white;
+            border-radius: 0.375rem;
+            box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+            max-width: 24rem;
+            width: 100%;
+            margin: 1.25rem;
+        }
+    </style>
 
     <title>Courses - Youdemy</title>
 </head>
@@ -49,10 +92,10 @@ echo "</pre>";
         <!-- start the project -->
         <!-- app layout -->
         <div id="app-layout" class="overflow-x-hidden flex">
-            <?php include '../../partials/navbar-vertical.php'; ?>
+            <?php include '../../components/navbar-vertical.php'; ?>
             <!-- app layout content -->
             <div id="app-layout-content" class="min-h-screen w-full min-w-[100vw] md:min-w-0 ml-[15.625rem] [transition:margin_0.25s_ease-out]">
-                <?php include '../../partials/top-navbar.php'; ?>
+                <?php include '../../components/top-navbar.php'; ?>
 
                 <!-- Page Header -->
                 <div class="bg-indigo-600 px-6 pt-6 pb-8 h-20 flex justify-between items-center mb-8">
@@ -60,6 +103,9 @@ echo "</pre>";
                     <nav class="text-white text-sm">
                         <a href="../dashboard.php" class="hover:underline">Dashboard</a> / Courses
                     </nav>
+                    <?php if ($role === 'teacher') : ?>
+                        <a href="../forms/addCourse.php" class="btn bg-white text-gray-800 border-gray-600 hover:bg-gray-100 hover:text-gray-800 hover:border-gray-200 active:bg-gray-100 active:text-gray-800 active:border-gray-200 focus:outline-none focus:ring-4 focus:ring-indigo-300">Create New Course</a>
+                    <?php endif; ?>
                 </div>
 
                 <!-- Courses Table -->
@@ -88,39 +134,45 @@ echo "</pre>";
                                 </thead>
                                 <tbody class="divide-y ">
                                     <?php
-                                    $count = 1;
-                                    foreach ($courses as $course) :
-                                        $tags = $courseContr->getCourseTags(6);
-
+                                    if (!empty($courses)) :
+                                        $count = 1;
+                                        foreach ($courses as $course) :
+                                            $tags = $courseContr->getCourseTags($course['id']);
                                     ?>
-                                        <tr class="border-gray-300 border-b ">
-                                            <td class="py-3 px-6 text-left"><?= $count++ ?></td>
-                                            <td class="py-3 px-6 text-left"><?= $course['title'] ?></td>
-                                            <td class="py-3 px-6 text-left"><?= $catgory->getCategoryName($course['categoryId']) ?></td>
-                                            <td class="py-3 px-6 text-left">
-                                                <?php foreach ($tags as $tag) : ?>
-                                                    <span class="bg-blue-100 text-blue-800 text-sm font-medium px-2 py-1 rounded"><?= Tag::getTagName($tag['tagId']) ?></span>
-                                                <?php endforeach; ?>
-                                            </td>
-                                            <td class="py-3 px-6 text-left"><?= User::getUserById($course['teacherId'])[0]['fullName'] ?></td>
-                                            <td class="py-3 px-6 text-left"><?= $course['creationDate'] ?></td>
-                                            <td class="py-3 px-6 text-left">20</td>
-                                            <td>
-                                                <?php if ($course['status'] === 'draft' && $_SESSION['user']['role'] === 'admin') : ?>
-                                                    <form action="" method="POST">
-                                                        <input type="hidden" name="course_id" value="<?= $course['id'] ?>">
-                                                        <button type="submit" class="bg-green-100 text-green-800 text-sm font-medium px-2 py-1 rounded" name="accept-course">Accept Course</button>
-                                                    </form>
-                                                <?php else : ?>
-                                                    <span class="badge badge-success p-2"><?= htmlspecialchars($course['status']) ?></span>
-                                                <?php endif; ?>
-                                            </td>
-                                            <td class="py-3 px-6 text-left">
-                                                <button class="btn btn-sm bg-indigo-500 text-white">Read</button>
-                                                <button class="btn btn-sm bg-red-500 text-white">Delete</button>
-                                            </td>
-                                        </tr>
-                                    <?php endforeach; ?>
+                                            <tr class="border-gray-300 border-b ">
+                                                <td class="py-3 px-6 text-left"><?= $count++ ?></td>
+                                                <td class="py-3 px-6 text-left"><?= $course['title'] ?></td>
+                                                <td class="py-3 px-6 text-left"><?= $catgory->getCategoryName($course['categoryId']) ?></td>
+                                                <td class="py-3 px-6 text-left">
+                                                    <?php foreach ($tags as $tag) : ?>
+                                                        <span class="bg-blue-100 text-blue-800 text-sm font-medium px-2 py-1 rounded"><?= Tag::getTagName($tag['tagId']) ?></span>
+                                                    <?php endforeach; ?>
+                                                </td>
+                                                <td class="py-3 px-6 text-left"><?= User::getUserById($course['teacherId'])[0]['fullName'] ?></td>
+                                                <td class="py-3 px-6 text-left"><?= $course['creationDate'] ?></td>
+                                                <td class="py-3 px-6 text-left">20</td>
+                                                <td>
+                                                    <?php if ($course['status'] === 'draft' && Session::get('user')[0]['role'] === 'admin') : ?>
+                                                        <form action="" method="POST">
+                                                            <input type="hidden" name="course_id" value="<?= $course['id'] ?>">
+                                                            <button type="submit" class="bg-yellow-100 text-yellow-800 text-sm font-medium px-2 py-1 rounded" name="accept-course">Accept Course</button>
+                                                        </form>
+                                                    <?php else : ?>
+                                                        <span class="bg-green-100 text-green-800 text-sm font-medium px-2 py-1 rounded" name="accept-course"><?= htmlspecialchars($course['status']) ?></span>
+                                                    <?php endif; ?>
+                                                </td>
+                                                <td class="py-3 px-6 text-left">
+                                                    <button class="btn btn-sm bg-indigo-500 text-white">Read</button>
+                                                    <?php if (Session::get('user')[0]['role'] === 'teacher') : ?>
+                                                        <a href="../forms/editCourse.php?courseId=<?= $course['id'] ?>" class="btn btn-sm bg-indigo-500 text-white">Edit</a>
+                                                    <?php endif; ?>
+                                                    <button class="btn btn-sm bg-red-500 text-white" onclick="openDeleteCourseModal(<?= $course['id'] ?>)">Delete</button>
+                                                </td>
+                                            </tr>
+                                        <?php endforeach; ?>
+                                    <?php else : ?>
+                                        <td class="py-3 px-6">No Courses Available</td>
+                                    <?php endif; ?>
                                 </tbody>
                             </table>
                         </div>
@@ -129,7 +181,10 @@ echo "</pre>";
             </div>
         </div>
         <!-- end of project -->
+
+        <?php include '../../components/modals.php' ?>
     </main>
+    <script src="../../../public/assets/js/main.js"></script>
     <script src="../../../public/assets/js/index.js"></script>
 
 </body>
